Checkers

+++ apps/api/views/image.py
@@ -8,7 +8,7 @@ from apps.api.errors import ProblemDetailException
 from apps.core.models import Image
 from apps.api.views.base import SecuredView
 from apps.api.response import SingleResponse
-
+from object_checker.base_object_checker import has_object_permission
 
 
 class ImageDetail(SecuredView):
@@ -34,8 +34,11 @@ class ImageDetail(SecuredView):
 
         image = self._get_image(request, image_id)
 
+        if not has_object_permission('check_delete', user=request.user, obj=image):
+            raise ProblemDetailException(request, _('Permission denied.'), status=HTTPStatus.FORBIDDEN)
+
         image.path.delete()
         image.hard_delete()
 
-        return SingleResponse(request)
+        return SingleResponse(request, status=HTTPStatus.NO_CONTENT)
